process-change: loosen requirements for approvers (#1124)
* process-change: loosen requirements for approvers

The approver role can be difficult to assign for repositories with lower
velocity or for those with a small number of (e.g. 2) primary
contributors, where the contributor seeking the approval role is making
most of the code contributions. This change attempts to address that
situation by removing a hard number requirement, and allowing for
individual contributions to be considered as well. Finally, instead of
simply suggesting that the contributor be in the OWNERS file, this
change strengthens that as a requirement.

Signed-off-by: Lance Ball <lball@redhat.com>

* squash: also loosen requirements for WG leads

Based on community feedback to changes in requirements for the Approver
role, this change also loosens the requirements for working group leads,
eliminating the hard number requirement for contributions.

Signed-off-by: Lance Ball <lball@redhat.com>