update ROLES.md to more accurately reflect reality (#1343)
* update ROLES.md to more accurately reflect reality

This is a first pass at rewriting the ROLES.md document to more
accurately reflect the reality of the roles and responsibilities in the
knative and knative-sandbox GitHub organizations. I have based these
changes, in part, on the CNCF Ladder template found here:

https://github.com/cncf/project-template/blob/main/CONTRIBUTOR_LADDER.md

Major changes include the addition of a Contributor role, which grants
no explicit privileges, combination of the Technical and Execution lead
roles for Working Groups into a single Working Group Lead role, and the
removal of the Scribe role, which appears to be more aspirational than
anything else.

I have removed the Requirements, Privileges, and Scope sections from the
summary table so that it reflects simply the Responsiblities of each
role, with details spelled out in the subsequent text.

The Responsibilities sections in the text could be further refined.

I also found it quite difficult to map the peribolos groups that are
reflected in `OWNERS` files to the Member and Approver Roles and would
appreciate a review of these items specifically. It is not clear, based
on what I can see in `./peribolos/knative.yaml`, what groups actually
allow for reviews and which allow for approvals and write permissions
on the repositories.

Signed-off-by: Lance Ball <lball@redhat.com>

* fixup: adjust NOTES text

Signed-off-by: Lance Ball <lball@redhat.com>

* fixup: apply suggestions from code review

Co-authored-by: Evan Anderson <evan.k.anderson@gmail.com>

* fixup: additional review feedback incorporated

Signed-off-by: Lance Ball <lball@redhat.com>

* fixup: additional review feedback incorporated

Signed-off-by: Lance Ball <lball@redhat.com>

* Apply suggestions from code review

Co-authored-by: Dawn Foster <geekygirldawn@gmail.com>

* fixup: remove election wording

Signed-off-by: Lance Ball <lball@redhat.com>

---------

Signed-off-by: Lance Ball <lball@redhat.com>
Co-authored-by: Evan Anderson <evan.k.anderson@gmail.com>
Co-authored-by: Dawn Foster <geekygirldawn@gmail.com>