[PROCESS CHANGE]: add a simple inactivity measure and process for removal (#1390)
* docs: add a simple inactivity measure and process for removal

The ROLES.md document was recently updated to more accurately reflect
how the Knative organization should work with regards to a "contribution
ladder" and methods of moving up and down that latter based on project
contributions. What was not addressed in those changes was how to handle
stepping away from the project or being removed for some reason.

This change is meant to address that omission. The text is
taken from the CNCF contribution ladder template located here:

https://github.com/cncf/project-template/blob/main/CONTRIBUTOR_LADDER.md#inactivity

When members step away from the project or step away from leadership roles, they
should be removed from teams that have elevated priviledges (writers,
reviewers, admin, etc).

When members choose to leave the project entirely due to changing
employment of life events, they should be removed from the knative and
knative-sandbox

When members fail to contribute for a period of one year, they may be
involuntarily removed from the knative and knative-sandbox GitHub organizations.

Signed-off-by: Lance Ball <lball@redhat.com>

* Update ROLES.md

Co-authored-by: Ali Ok <aliok@redhat.com>

---------

Signed-off-by: Lance Ball <lball@redhat.com>
Co-authored-by: Carlos Santana <csantana23@gmail.com>
Co-authored-by: Ali Ok <aliok@redhat.com>